Initial Project
Forgot to make a git repo until I was a little bit into the project
because I needed to move the code to another computer, but
not much is actually set up.
I created a very basic login page and created the fragment for a very
basic feed page.
I added the Navigation component for hooking up the two.