Basic Login
Wired up the login fragment to launch off a call to the backend to authenticate the provided email and password.
There is a lot of stuff missing. The biggest thing is error states, like if you don't include an email or password or if the request fails.
Another thing is I don't like using callbacks for every network request, that's going to get old really fast. I haven't decided if I want to go full Observables/RxJava yet so leaving as Callback now.