Paging!
I couldn't seem to figure out Google's Paging library DataSources so I made my own basic version of a paging implementation.
As the RecyclerView scrolls, it tracks how far it has scrolled. Once it passes 75%, it tries to load the next set of data.
When new data comes in, it is added to the data in the Adapter and the data set is updated.